Round 1 Edits
(1) Rename it Tool for Intelligibility Assessment (I don't expect this to reach acronym-level usage, but we want to avoid IAT since that belongs to something else)
(2) Edit "Sentences" to "SIT Sentences" in both the table header and column header
(3) Allow user to sort columns
(4) Have the results return a table with all sentence/listener combos that meet the user-specified accuracy level or greater. We want them to be able to see whether adding sentences or listeners gives them a boost in accuracy that may be worthwhile for their purposes.